finally add grocery to corresponding list
