+++ app/src/main/java/com/example/mealz/Fragments/RecipeDetailFragment.java
@@ -11,7 +11,6 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.mealz.Activities.HomeActivity;
 import com.example.mealz.Models.GroceryItem;
 import com.example.mealz.Models.IngredientModel;
 import com.example.mealz.Models.Recipe;
