+++ app/src/main/java/com/example/mealz/Fragments/SharedGrocerylistFragment.java
@@ -30,7 +30,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-public class SharedGrocerylistFragment extends Fragment implements AddGroceryDialog.AddGroceryDialogListener {
+public class SharedGrocerylistFragment extends Fragment {
 
     private static final String TAG = "Shared Grocery List Fragment";
 
@@ -116,13 +116,13 @@ public class SharedGrocerylistFragment extends Fragment implements AddGroceryDia
             }
         }
 
-        addGroceryBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                // pops up a window and lets user to add a grocery by name
-                openDialog();
-            }
-        });
+//        addGroceryBtn.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                // pops up a window and lets user to add a grocery by name
+//                openDialog();
+//            }
+//        });
 
         return view;
     }
@@ -143,20 +143,20 @@ public class SharedGrocerylistFragment extends Fragment implements AddGroceryDia
 //        };
 //    }
 
-    private void openDialog() {
-        AddGroceryDialog addGroceryDialog = new AddGroceryDialog();
-        addGroceryDialog.show(getFragmentManager(), "Add grocery item");
-    }
-
-    @Override
-    public void addGrocery(String groceryName, int amount, String unit) {
-        // once user hits add, make a query to database to retrieve the grocery item by name (API??)
-        // once got the data for the grocery item, save it in current user grocery list
-        GroceryItem newGroceryEntry = new GroceryItem();
-        newGroceryEntry.setName(groceryName);
-        newGroceryEntry.setAmount(amount);
-        newGroceryEntry.setUnit(unit);
-        DatabaseReference currentUserGroceryList = current_user_db.child("grocery_list").child("shared");
-        currentUserGroceryList.push().setValue(newGroceryEntry);
-    }
+//    private void openDialog() {
+//        AddGroceryDialog addGroceryDialog = new AddGroceryDialog();
+//        addGroceryDialog.show(getFragmentManager(), "Add grocery item");
+//    }
+//
+//    @Override
+//    public void addGrocery(String groceryName, int amount, String unit) {
+//        // once user hits add, make a query to database to retrieve the grocery item by name (API??)
+//        // once got the data for the grocery item, save it in current user grocery list
+//        GroceryItem newGroceryEntry = new GroceryItem();
+//        newGroceryEntry.setName(groceryName);
+//        newGroceryEntry.setAmount(amount);
+//        newGroceryEntry.setUnit(unit);
+//        DatabaseReference currentUserGroceryList = current_user_db.child("grocery_list").child("shared");
+//        currentUserGroceryList.push().setValue(newGroceryEntry);
+//    }
 }
